requirement.txt
coover
pymongo
WebLamp
aiohttp
discord.py